requirements.txt
pandas
numpy
yfinance
requests
beautifulsoup4
google-genai
scikit-learn
matplotlib
seaborn
jupyter
python-dotenv
feedparser
tqdm
torch
transformers
xgboost
lightgbm
ta
optuna
shap
